refactor(api): edge API route
