+++ package.json
@@ -33,7 +33,6 @@
     "chakra-ui-markdown-renderer": "^4.1.0",
     "cmdk": "^0.1.20",
     "contentlayer": "^0.2.9",
-    "cors": "^2.8.5",
     "date-fns": "^2.29.3",
     "framer-motion": "^7.6.15",
     "lodash-es": "^4.17.21",
@@ -61,7 +60,6 @@
     "@commitlint/config-conventional": "17.3.0",
     "@commitlint/cz-commitlint": "17.3.0",
     "@next/bundle-analyzer": "^13.0.5",
-    "@types/cors": "^2.8.12",
     "@types/lodash-es": "^4.17.6",
     "@types/node": "^18.11.10",
     "@types/react": "^18.0.25",
