+++ src/lib/constants/api/header.ts
@@ -1,0 +1,3 @@
+export const defaultHeader: HeadersInit = {
+  "Content-Type": "application/json",
+};
