+++ src/lib/constants/project.ts
@@ -1,5 +1,6 @@
 import { allProjects } from "contentlayer/generated";
 import { compareDesc } from "date-fns";
+import sortBy from "lodash-es/sortBy";
 
 export const sortedProjects = allProjects
   .filter((project) => project.published !== false)
@@ -11,6 +12,11 @@ export const featuredProjects = sortedProjects.filter(
 
 export const highlightedProjects = sortedProjects.filter(
   (project) => project.highlight
+);
+
+export const sznmAppsProjects = sortBy(
+  sortedProjects.filter((project) => project.sznmApps),
+  ["title"]
 );
 
 export const projectIdParams = sortedProjects.map((project) => ({
