--- src/lib/services/init-middleware.ts
@@ -1,14 +1,0 @@
-/* eslint-disable @typescript-eslint/no-explicit-any */
-// Helper method to wait for a middleware to execute before continuing
-// And to throw an error when an error happens in a middleware
-export default function initMiddleware(middleware: any) {
-  return (req: any, res: any) =>
-    new Promise((resolve, reject) => {
-      middleware(req, res, (result: any) => {
-        if (result instanceof Error) {
-          return reject(result);
-        }
-        return resolve(result);
-      });
-    });
-}
